+++ crates/omnix-ci/src/github/mod.rs
@@ -1,3 +1,4 @@
 //! GitHub related types and functions.
+pub mod actions;
 pub mod matrix;
 pub mod pull_request;
